fix type

+++ tcp/tcpv4_windows_test.go
@@ -96,9 +96,9 @@ func TestSendAndReceive(t *testing.T) {
 			require.NoError(t, err)
 			assert.Empty(t, cmp.Diff(test.expected, actual, cmpopts.IgnoreFields(result.TracerouteHop{}, "RTT")))
 			if !test.mockHopIP.Equal(net.IP{}) { // only if we get a hop IP back should RTT be >0
-				assert.Greater(t, actual.RTT, 0)
+				assert.Greater(t, actual.RTT, float64(0))
 			} else {
-				assert.Equal(t, actual.RTT, 0)
+				assert.Equal(t, actual.RTT, float64(0))
 			}
 		})
 	}
